Continue dioxus cleanup
