+++ mlm_web_dioxus/src/components/status_message.rs
@@ -6,17 +6,18 @@ pub fn StatusMessage(mut status_msg: Signal<Option<(String, bool)>>) -> Element 
         return rsx! {};
     };
 
+    let class = if is_error {
+        "status-message error"
+    } else {
+        "status-message success"
+    };
+
     rsx! {
-        div {
-            class: if is_error { "error" } else { "success" },
-            style: if is_error {
-                "padding: 10px; margin-bottom: 10px; border-radius: 4px; color: #000; background: #fdd;"
-            } else {
-                "padding: 10px; margin-bottom: 10px; border-radius: 4px; color: #000; background: #dfd;"
-            },
+        div { class,
             "{msg}"
             button {
-                style: "margin-left: 10px; cursor: pointer;",
+                r#type: "button",
+                "aria-label": "Dismiss status message",
                 onclick: move |_| status_msg.set(None),
                 "⨯"
             }
