+++ mlm_web_dioxus/src/duplicate/components.rs
@@ -7,376 +7,9 @@ use crate::components::{
     parse_query_enum, set_location_query_string,
 };
 use dioxus::prelude::*;
-use serde::{Deserialize, Serialize};
 
-#[cfg(feature = "server")]
-use crate::error::{IntoServerFnError, OptionIntoServerFnError};
-#[cfg(feature = "server")]
-use crate::utils::format_timestamp_db;
-#[cfg(feature = "server")]
-use mlm_core::{ContextExt, Torrent, cleaner::clean_torrent};
-#[cfg(feature = "server")]
-use mlm_db::{DatabaseExt as _, DuplicateTorrent, SelectedTorrent, Timestamp, TorrentCost, ids};
-#[cfg(feature = "server")]
-use mlm_parse::normalize_title;
-
-#[derive(Clone, Copy, PartialEq, Serialize, Deserialize, Debug)]
-#[serde(rename_all = "lowercase")]
-pub enum DuplicatePageSort {
-    Kind,
-    Title,
-    Authors,
-    Narrators,
-    Series,
-    Size,
-    CreatedAt,
-}
-
-#[derive(Clone, Copy, PartialEq, Eq, Serialize, Deserialize, Debug)]
-#[serde(rename_all = "snake_case")]
-pub enum DuplicatePageFilter {
-    Kind,
-    Title,
-    Author,
-    Narrator,
-    Series,
-    Filetype,
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
-pub struct DuplicateSeries {
-    pub name: String,
-    pub entries: String,
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
-pub struct DuplicateMeta {
-    pub title: String,
-    pub media_type: String,
-    pub authors: Vec<String>,
-    pub narrators: Vec<String>,
-    pub series: Vec<DuplicateSeries>,
-    pub size: String,
-    pub filetypes: Vec<String>,
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
-pub struct DuplicateCandidateRow {
-    pub mam_id: u64,
-    pub meta: DuplicateMeta,
-    pub created_at: String,
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
-pub struct DuplicateOriginalRow {
-    pub id: String,
-    pub mam_id: Option<u64>,
-    pub meta: DuplicateMeta,
-    pub linked: bool,
-    pub linked_path: Option<String>,
-    pub created_at: String,
-    pub abs_id: Option<String>,
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
-pub struct DuplicatePairRow {
-    pub torrent: DuplicateCandidateRow,
-    pub duplicate_of: DuplicateOriginalRow,
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, Default, PartialEq)]
-pub struct DuplicateData {
-    pub torrents: Vec<DuplicatePairRow>,
-    pub total: usize,
-    pub from: usize,
-    pub page_size: usize,
-    pub abs_url: Option<String>,
-}
-
-#[derive(Clone, Copy, PartialEq, Serialize, Deserialize, Debug)]
-#[serde(rename_all = "snake_case")]
-pub enum DuplicateBulkAction {
-    Replace,
-    Remove,
-}
-
-impl DuplicateBulkAction {
-    fn label(self) -> &'static str {
-        match self {
-            Self::Replace => "replace original",
-            Self::Remove => "remove duplicate",
-        }
-    }
-
-    fn success_label(self) -> &'static str {
-        match self {
-            Self::Replace => "Replaced original torrents",
-            Self::Remove => "Removed duplicate torrents",
-        }
-    }
-}
-
-#[cfg(feature = "server")]
-fn matches_filter(t: &DuplicateTorrent, field: DuplicatePageFilter, value: &str) -> bool {
-    match field {
-        DuplicatePageFilter::Kind => t.meta.media_type.as_str() == value,
-        DuplicatePageFilter::Title => t.meta.title == value,
-        DuplicatePageFilter::Author => t.meta.authors.contains(&value.to_string()),
-        DuplicatePageFilter::Narrator => t.meta.narrators.contains(&value.to_string()),
-        DuplicatePageFilter::Series => t.meta.series.iter().any(|s| s.name == value),
-        DuplicatePageFilter::Filetype => t.meta.filetypes.iter().any(|f| f == value),
-    }
-}
-
-#[cfg(feature = "server")]
-fn convert_candidate_row(t: &DuplicateTorrent) -> DuplicateCandidateRow {
-    DuplicateCandidateRow {
-        mam_id: t.mam_id,
-        meta: DuplicateMeta {
-            title: t.meta.title.clone(),
-            media_type: t.meta.media_type.as_str().to_string(),
-            authors: t.meta.authors.clone(),
-            narrators: t.meta.narrators.clone(),
-            series: t
-                .meta
-                .series
-                .iter()
-                .map(|series| DuplicateSeries {
-                    name: series.name.clone(),
-                    entries: series.entries.to_string(),
-                })
-                .collect(),
-            size: t.meta.size.to_string(),
-            filetypes: t.meta.filetypes.clone(),
-        },
-        created_at: format_timestamp_db(&t.created_at),
-    }
-}
-
-#[cfg(feature = "server")]
-fn convert_original_row(t: &Torrent) -> DuplicateOriginalRow {
-    DuplicateOriginalRow {
-        id: t.id.clone(),
-        mam_id: t.mam_id,
-        meta: DuplicateMeta {
-            title: t.meta.title.clone(),
-            media_type: t.meta.media_type.as_str().to_string(),
-            authors: t.meta.authors.clone(),
-            narrators: t.meta.narrators.clone(),
-            series: t
-                .meta
-                .series
-                .iter()
-                .map(|series| DuplicateSeries {
-                    name: series.name.clone(),
-                    entries: series.entries.to_string(),
-                })
-                .collect(),
-            size: t.meta.size.to_string(),
-            filetypes: t.meta.filetypes.clone(),
-        },
-        linked: t.library_path.is_some(),
-        linked_path: t
-            .library_path
-            .as_ref()
-            .map(|path| path.to_string_lossy().to_string()),
-        created_at: format_timestamp_db(&t.created_at),
-        abs_id: t.meta.ids.get(ids::ABS).cloned(),
-    }
-}
-
-#[server]
-pub async fn get_duplicate_data(
-    sort: Option<DuplicatePageSort>,
-    asc: bool,
-    filters: Vec<(DuplicatePageFilter, String)>,
-    from: Option<usize>,
-    page_size: Option<usize>,
-) -> Result<DuplicateData, ServerFnError> {
-    let context = crate::error::get_context()?;
-
-    let mut from_val = from.unwrap_or(0);
-    let page_size_val = page_size.unwrap_or(500);
-
-    let r = context.db().r_transaction().server_err()?;
-
-    let mut duplicates = r
-        .scan()
-        .primary::<DuplicateTorrent>()
-        .server_err()?
-        .all()
-        .server_err()?
-        .filter_map(Result::ok)
-        .filter(|t| {
-            filters
-                .iter()
-                .all(|(field, value)| matches_filter(t, *field, value))
-        })
-        .collect::<Vec<_>>();
-
-    if let Some(sort_by) = sort {
-        duplicates.sort_by(|a, b| {
-            let ord = match sort_by {
-                DuplicatePageSort::Kind => a.meta.media_type.cmp(&b.meta.media_type),
-                DuplicatePageSort::Title => a.meta.title.cmp(&b.meta.title),
-                DuplicatePageSort::Authors => a.meta.authors.cmp(&b.meta.authors),
-                DuplicatePageSort::Narrators => a.meta.narrators.cmp(&b.meta.narrators),
-                DuplicatePageSort::Series => a.meta.series.cmp(&b.meta.series),
-                DuplicatePageSort::Size => a.meta.size.cmp(&b.meta.size),
-                DuplicatePageSort::CreatedAt => a.created_at.cmp(&b.created_at),
-            };
-            if asc { ord.reverse() } else { ord }
-        });
-    }
-
-    let total = duplicates.len();
-    if page_size_val > 0 && from_val >= total && total > 0 {
-        from_val = ((total - 1) / page_size_val) * page_size_val;
-    }
-
-    let limit = if page_size_val == 0 {
-        usize::MAX
-    } else {
-        page_size_val
-    };
-
-    let mut rows = Vec::new();
-    for duplicate in duplicates.into_iter().skip(from_val).take(limit) {
-        let Some(duplicate_of_id) = &duplicate.duplicate_of else {
-            continue;
-        };
-        let Some(duplicate_of) = r
-            .get()
-            .primary::<Torrent>(duplicate_of_id.clone())
-            .server_err()?
-        else {
-            continue;
-        };
-        rows.push(DuplicatePairRow {
-            torrent: convert_candidate_row(&duplicate),
-            duplicate_of: convert_original_row(&duplicate_of),
-        });
-    }
-
-    let abs_url = context
-        .config()
-        .await
-        .audiobookshelf
-        .as_ref()
-        .map(|abs| abs.url.clone());
-
-    Ok(DuplicateData {
-        torrents: rows,
-        total,
-        from: from_val,
-        page_size: page_size_val,
-        abs_url,
-    })
-}
-
-#[server]
-pub async fn apply_duplicate_action(
-    action: DuplicateBulkAction,
-    torrent_ids: Vec<u64>,
-) -> Result<(), ServerFnError> {
-    if torrent_ids.is_empty() {
-        return Err(ServerFnError::new("No torrents selected"));
-    }
-
-    let context = crate::error::get_context()?;
-    let config = context.config().await;
-
-    match action {
-        DuplicateBulkAction::Replace => {
-            let mam = context.mam().server_err()?;
-            for mam_id in torrent_ids {
-                let r = context.db().r_transaction().server_err()?;
-                let Some(duplicate_torrent) =
-                    r.get().primary::<DuplicateTorrent>(mam_id).server_err()?
-                else {
-                    continue;
-                };
-                let Some(hash) = duplicate_torrent.duplicate_of.clone() else {
-                    return Err(ServerFnError::new("No duplicate_of set"));
-                };
-                let Some(duplicate_of) = r.get().primary::<Torrent>(hash).server_err()? else {
-                    return Err(ServerFnError::new("Could not find original torrent"));
-                };
-
-                let Some(mam_torrent) = mam
-                    .get_torrent_info_by_id(duplicate_torrent.mam_id)
-                    .await
-                    .server_err()?
-                else {
-                    return Err(ServerFnError::new(
-                        "Could not find duplicate torrent on MaM",
-                    ));
-                };
-
-                let meta = mam_torrent.as_meta().server_err()?;
-                let title_search = normalize_title(&meta.title);
-                let tags: Vec<_> = config
-                    .tags
-                    .iter()
-                    .filter(|t| t.filter.matches(&mam_torrent))
-                    .collect();
-                let category = tags.iter().find_map(|t| t.category.clone());
-                let tags = tags.iter().flat_map(|t| t.tags.clone()).collect();
-                let cost = if mam_torrent.vip {
-                    TorrentCost::Vip
-                } else if mam_torrent.personal_freeleech {
-                    TorrentCost::PersonalFreeleech
-                } else if mam_torrent.free {
-                    TorrentCost::GlobalFreeleech
-                } else {
-                    TorrentCost::TryWedge
-                };
-
-                let (_guard, rw) = context.db().rw_async().await.server_err()?;
-                rw.insert(SelectedTorrent {
-                    mam_id: mam_torrent.id,
-                    hash: None,
-                    dl_link: mam_torrent
-                        .dl
-                        .clone()
-                        .or_else(|| duplicate_torrent.dl_link.clone())
-                        .ok_or_server_err("No download link for duplicate torrent")?,
-                    unsat_buffer: None,
-                    wedge_buffer: None,
-                    cost,
-                    category,
-                    tags,
-                    title_search,
-                    meta,
-                    grabber: None,
-                    created_at: Timestamp::now(),
-                    started_at: None,
-                    removed_at: None,
-                })
-                .server_err()?;
-                rw.remove(duplicate_torrent).server_err()?;
-                rw.commit().server_err()?;
-
-                clean_torrent(&config, context.db(), duplicate_of, false, &context.events)
-                    .await
-                    .server_err()?;
-            }
-        }
-        DuplicateBulkAction::Remove => {
-            let (_guard, rw) = context.db().rw_async().await.server_err()?;
-            for mam_id in torrent_ids {
-                let Some(torrent) = rw.get().primary::<DuplicateTorrent>(mam_id).server_err()?
-                else {
-                    continue;
-                };
-                rw.remove(torrent).server_err()?;
-            }
-            rw.commit().server_err()?;
-        }
-    }
-
-    Ok(())
-}
+use super::server_fns::{apply_duplicate_action, get_duplicate_data};
+use super::types::*;
 
 fn filter_name(filter: DuplicatePageFilter) -> &'static str {
     match filter {
@@ -526,11 +159,15 @@ pub fn DuplicatePage() -> Element {
             let mut filters_signal = filters;
             let mut from = from;
             let mut page_size = page_size;
+            let mut selected = selected;
+            let mut last_selected_idx = last_selected_idx;
             sort.set(route_state.sort);
             asc.set(route_state.asc);
             filters_signal.set(route_state.filters);
             from.set(route_state.from);
             page_size.set(route_state.page_size);
+            selected.set(BTreeSet::new());
+            last_selected_idx.set(None);
             last_request_key.set(route_request_key);
             if let Some(resource) = duplicate_data.as_mut() {
                 resource.restart();
@@ -538,12 +175,15 @@ pub fn DuplicatePage() -> Element {
         }
     }
 
-    if let Some(value) = &value {
-        let value = value.read();
-        if let Some(Ok(data)) = &*value {
-            cached.set(Some(data.clone()));
+    let cache_value = value;
+    use_effect(move || {
+        if let Some(value) = &cache_value {
+            let value = value.read();
+            if let Some(Ok(data)) = &*value {
+                cached.set(Some(data.clone()));
+            }
         }
-    }
+    });
 
     let data_to_show = {
         if let Some(value) = &value {
