+++ mlm_web_dioxus/src/duplicate/mod.rs
@@ -1,0 +1,7 @@
+mod components;
+mod server_fns;
+mod types;
+
+pub use components::DuplicatePage;
+pub use server_fns::{apply_duplicate_action, get_duplicate_data};
+pub use types::*;
