+++ mlm_web_dioxus/src/events/mod.rs
@@ -4,7 +4,7 @@ mod types;
 
 pub use components::{EventContent, EventListItem, EventsPage};
 pub use server_fns::get_events_data;
-pub use types::{EventData, EventWithTorrentData};
+pub use types::{EventData, EventWithTorrentData, EventsFilter};
 
 // Re-export the SSE trigger for backward compatibility
 pub use crate::sse::EVENTS_UPDATE_TRIGGER;
