+++ mlm_web_dioxus/src/replaced/components.rs
@@ -4,389 +4,11 @@ use crate::components::{
     encode_query_enum, parse_location_query_pairs, parse_query_enum, set_location_query_string,
 };
 use dioxus::prelude::*;
-use serde::{Deserialize, Serialize};
 use std::collections::BTreeSet;
-#[cfg(feature = "server")]
-use std::str::FromStr;
 use std::sync::Arc;
 
-#[cfg(feature = "server")]
-use crate::error::IntoServerFnError;
-#[cfg(feature = "server")]
-use crate::utils::format_timestamp_db;
-#[cfg(feature = "server")]
-use mlm_core::{
-    ContextExt, Torrent,
-    linker::{refresh_mam_metadata, refresh_metadata_relink},
-};
-#[cfg(feature = "server")]
-use mlm_db::{DatabaseExt as _, Language, TorrentKey, ids};
-
-#[derive(Clone, Copy, PartialEq, Serialize, Deserialize, Debug)]
-#[serde(rename_all = "lowercase")]
-pub enum ReplacedPageSort {
-    Kind,
-    Title,
-    Authors,
-    Narrators,
-    Series,
-    Language,
-    Size,
-    Replaced,
-    CreatedAt,
-}
-
-#[derive(Clone, Copy, PartialEq, Eq, Serialize, Deserialize, Debug)]
-#[serde(rename_all = "snake_case")]
-pub enum ReplacedPageFilter {
-    Kind,
-    Title,
-    Author,
-    Narrator,
-    Series,
-    Language,
-    Filetype,
-    Linked,
-}
-
-#[derive(Clone, Copy, Debug, Serialize, Deserialize, PartialEq)]
-pub struct ReplacedPageColumns {
-    pub authors: bool,
-    pub narrators: bool,
-    pub series: bool,
-    pub language: bool,
-    pub size: bool,
-    pub filetypes: bool,
-}
-
-impl Default for ReplacedPageColumns {
-    fn default() -> Self {
-        Self {
-            authors: true,
-            narrators: true,
-            series: true,
-            language: false,
-            size: true,
-            filetypes: true,
-        }
-    }
-}
-
-impl ReplacedPageColumns {
-    fn table_grid_template(self) -> String {
-        let mut cols = vec!["30px", "110px", "2fr"];
-        if self.authors {
-            cols.push("1fr");
-        }
-        if self.narrators {
-            cols.push("1fr");
-        }
-        if self.series {
-            cols.push("1fr");
-        }
-        if self.language {
-            cols.push("100px");
-        }
-        if self.size {
-            cols.push("81px");
-        }
-        if self.filetypes {
-            cols.push("100px");
-        }
-        cols.push("157px");
-        cols.push("157px");
-        cols.push("132px");
-        cols.join(" ")
-    }
-
-    pub fn get(self, col: ReplacedColumn) -> bool {
-        match col {
-            ReplacedColumn::Authors => self.authors,
-            ReplacedColumn::Narrators => self.narrators,
-            ReplacedColumn::Series => self.series,
-            ReplacedColumn::Language => self.language,
-            ReplacedColumn::Size => self.size,
-            ReplacedColumn::Filetypes => self.filetypes,
-        }
-    }
-
-    pub fn set(&mut self, col: ReplacedColumn, enabled: bool) {
-        match col {
-            ReplacedColumn::Authors => self.authors = enabled,
-            ReplacedColumn::Narrators => self.narrators = enabled,
-            ReplacedColumn::Series => self.series = enabled,
-            ReplacedColumn::Language => self.language = enabled,
-            ReplacedColumn::Size => self.size = enabled,
-            ReplacedColumn::Filetypes => self.filetypes = enabled,
-        }
-    }
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
-pub struct ReplacedSeries {
-    pub name: String,
-    pub entries: String,
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
-pub struct ReplacedMeta {
-    pub title: String,
-    pub media_type: String,
-    pub authors: Vec<String>,
-    pub narrators: Vec<String>,
-    pub series: Vec<ReplacedSeries>,
-    pub language: Option<String>,
-    pub size: String,
-    pub filetypes: Vec<String>,
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
-pub struct ReplacedRow {
-    pub id: String,
-    pub mam_id: Option<u64>,
-    pub meta: ReplacedMeta,
-    pub linked: bool,
-    pub created_at: String,
-    pub replaced_at: Option<String>,
-    pub abs_id: Option<String>,
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
-pub struct ReplacedPairRow {
-    pub torrent: ReplacedRow,
-    pub replacement: ReplacedRow,
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, Default, PartialEq)]
-pub struct ReplacedData {
-    pub torrents: Vec<ReplacedPairRow>,
-    pub total: usize,
-    pub from: usize,
-    pub page_size: usize,
-    pub abs_url: Option<String>,
-}
-
-#[derive(Clone, Copy, PartialEq, Serialize, Deserialize, Debug)]
-#[serde(rename_all = "snake_case")]
-pub enum ReplacedBulkAction {
-    Refresh,
-    RefreshRelink,
-    Remove,
-}
-
-impl ReplacedBulkAction {
-    fn label(self) -> &'static str {
-        match self {
-            Self::Refresh => "refresh metadata",
-            Self::RefreshRelink => "refresh metadata and relink",
-            Self::Remove => "remove torrent from MLM",
-        }
-    }
-
-    fn success_label(self) -> &'static str {
-        match self {
-            Self::Refresh => "Refreshed metadata",
-            Self::RefreshRelink => "Refreshed metadata and relinked",
-            Self::Remove => "Removed torrents",
-        }
-    }
-}
-
-#[cfg(feature = "server")]
-fn matches_filter(t: &Torrent, field: ReplacedPageFilter, value: &str) -> bool {
-    match field {
-        ReplacedPageFilter::Kind => t.meta.media_type.as_str() == value,
-        ReplacedPageFilter::Title => t.meta.title == value,
-        ReplacedPageFilter::Author => t.meta.authors.contains(&value.to_string()),
-        ReplacedPageFilter::Narrator => t.meta.narrators.contains(&value.to_string()),
-        ReplacedPageFilter::Series => t.meta.series.iter().any(|s| s.name == value),
-        ReplacedPageFilter::Language => {
-            if value.is_empty() {
-                t.meta.language.is_none()
-            } else {
-                t.meta.language == Language::from_str(value).ok()
-            }
-        }
-        ReplacedPageFilter::Filetype => t.meta.filetypes.iter().any(|f| f == value),
-        ReplacedPageFilter::Linked => t.library_path.is_some() == (value == "true"),
-    }
-}
-
-#[cfg(feature = "server")]
-fn convert_row(t: &Torrent) -> ReplacedRow {
-    ReplacedRow {
-        id: t.id.clone(),
-        mam_id: t.mam_id,
-        meta: ReplacedMeta {
-            title: t.meta.title.clone(),
-            media_type: t.meta.media_type.as_str().to_string(),
-            authors: t.meta.authors.clone(),
-            narrators: t.meta.narrators.clone(),
-            series: t
-                .meta
-                .series
-                .iter()
-                .map(|series| ReplacedSeries {
-                    name: series.name.clone(),
-                    entries: series.entries.to_string(),
-                })
-                .collect(),
-            language: t.meta.language.map(|l| l.to_str().to_string()),
-            size: t.meta.size.to_string(),
-            filetypes: t.meta.filetypes.clone(),
-        },
-        linked: t.library_path.is_some(),
-        created_at: format_timestamp_db(&t.created_at),
-        replaced_at: t
-            .replaced_with
-            .as_ref()
-            .map(|(_, ts)| format_timestamp_db(ts)),
-        abs_id: t.meta.ids.get(ids::ABS).cloned(),
-    }
-}
-
-#[server]
-pub async fn get_replaced_data(
-    sort: Option<ReplacedPageSort>,
-    asc: bool,
-    filters: Vec<(ReplacedPageFilter, String)>,
-    from: Option<usize>,
-    page_size: Option<usize>,
-    _show: ReplacedPageColumns,
-) -> Result<ReplacedData, ServerFnError> {
-    let context = crate::error::get_context()?;
-
-    let mut from_val = from.unwrap_or(0);
-    let page_size_val = page_size.unwrap_or(500);
-
-    let r = context.db().r_transaction().server_err()?;
-
-    let mut replaced = r
-        .scan()
-        .secondary::<Torrent>(TorrentKey::created_at)
-        .server_err()?
-        .all()
-        .server_err()?
-        .rev()
-        .filter_map(Result::ok)
-        .filter(|t| t.replaced_with.is_some())
-        .filter(|t| {
-            filters
-                .iter()
-                .all(|(field, value)| matches_filter(t, *field, value))
-        })
-        .collect::<Vec<_>>();
-
-    if let Some(sort_by) = sort {
-        replaced.sort_by(|a, b| {
-            let ord = match sort_by {
-                ReplacedPageSort::Kind => a.meta.media_type.cmp(&b.meta.media_type),
-                ReplacedPageSort::Title => a.meta.title.cmp(&b.meta.title),
-                ReplacedPageSort::Authors => a.meta.authors.cmp(&b.meta.authors),
-                ReplacedPageSort::Narrators => a.meta.narrators.cmp(&b.meta.narrators),
-                ReplacedPageSort::Series => a.meta.series.cmp(&b.meta.series),
-                ReplacedPageSort::Language => a.meta.language.cmp(&b.meta.language),
-                ReplacedPageSort::Size => a.meta.size.cmp(&b.meta.size),
-                ReplacedPageSort::Replaced => a
-                    .replaced_with
-                    .as_ref()
-                    .map(|r| r.1)
-                    .cmp(&b.replaced_with.as_ref().map(|r| r.1)),
-                ReplacedPageSort::CreatedAt => a.created_at.cmp(&b.created_at),
-            };
-            if asc { ord.reverse() } else { ord }
-        });
-    }
-
-    let total = replaced.len();
-    if page_size_val > 0 && from_val >= total && total > 0 {
-        from_val = ((total - 1) / page_size_val) * page_size_val;
-    }
-
-    let limit = if page_size_val == 0 {
-        usize::MAX
-    } else {
-        page_size_val
-    };
-
-    let mut rows = Vec::new();
-    for torrent in replaced.into_iter().skip(from_val).take(limit) {
-        let Some((replacement_id, _)) = &torrent.replaced_with else {
-            continue;
-        };
-        let Some(replacement) = r
-            .get()
-            .primary::<Torrent>(replacement_id.clone())
-            .server_err()?
-        else {
-            continue;
-        };
-        rows.push(ReplacedPairRow {
-            torrent: convert_row(&torrent),
-            replacement: convert_row(&replacement),
-        });
-    }
-
-    let abs_url = context
-        .config()
-        .await
-        .audiobookshelf
-        .as_ref()
-        .map(|abs| abs.url.clone());
-
-    Ok(ReplacedData {
-        torrents: rows,
-        total,
-        from: from_val,
-        page_size: page_size_val,
-        abs_url,
-    })
-}
-
-#[server]
-pub async fn apply_replaced_action(
-    action: ReplacedBulkAction,
-    torrent_ids: Vec<String>,
-) -> Result<(), ServerFnError> {
-    if torrent_ids.is_empty() {
-        return Err(ServerFnError::new("No torrents selected"));
-    }
-
-    let context = crate::error::get_context()?;
-
-    match action {
-        ReplacedBulkAction::Refresh => {
-            let config = context.config().await;
-            let mam = context.mam().server_err()?;
-            for id in torrent_ids {
-                refresh_mam_metadata(&config, context.db(), &mam, id, &context.events)
-                    .await
-                    .server_err()?;
-            }
-        }
-        ReplacedBulkAction::RefreshRelink => {
-            let config = context.config().await;
-            let mam = context.mam().server_err()?;
-            for id in torrent_ids {
-                refresh_metadata_relink(&config, context.db(), &mam, id, &context.events)
-                    .await
-                    .server_err()?;
-            }
-        }
-        ReplacedBulkAction::Remove => {
-            let (_guard, rw) = context.db().rw_async().await.server_err()?;
-            for id in torrent_ids {
-                let Some(torrent) = rw.get().primary::<Torrent>(id).server_err()? else {
-                    continue;
-                };
-                rw.remove(torrent).server_err()?;
-            }
-            rw.commit().server_err()?;
-        }
-    }
-
-    Ok(())
-}
+use super::server_fns::*;
+use super::types::*;
 
 fn filter_name(filter: ReplacedPageFilter) -> &'static str {
     match filter {
@@ -398,54 +20,6 @@ fn filter_name(filter: ReplacedPageFilter) -> &'static str {
         ReplacedPageFilter::Language => "Language",
         ReplacedPageFilter::Filetype => "Filetypes",
         ReplacedPageFilter::Linked => "Linked",
-    }
-}
-
-impl PageColumns for ReplacedPageColumns {
-    fn to_query_value(&self) -> String {
-        let mut values = Vec::new();
-        if self.authors {
-            values.push("author");
-        }
-        if self.narrators {
-            values.push("narrator");
-        }
-        if self.series {
-            values.push("series");
-        }
-        if self.language {
-            values.push("language");
-        }
-        if self.size {
-            values.push("size");
-        }
-        if self.filetypes {
-            values.push("filetype");
-        }
-        values.join(",")
-    }
-
-    fn from_query_value(value: &str) -> Self {
-        let mut show = ReplacedPageColumns {
-            authors: false,
-            narrators: false,
-            series: false,
-            language: false,
-            size: false,
-            filetypes: false,
-        };
-        for item in value.split(',') {
-            match item {
-                "author" => show.authors = true,
-                "narrator" => show.narrators = true,
-                "series" => show.series = true,
-                "language" => show.language = true,
-                "size" => show.size = true,
-                "filetype" => show.filetypes = true,
-                _ => {}
-            }
-        }
-        show
     }
 }
 
@@ -531,25 +105,6 @@ fn build_query_url(
     build_query_string(&params)
 }
 
-#[derive(Clone, Copy)]
-enum ReplacedColumn {
-    Authors,
-    Narrators,
-    Series,
-    Language,
-    Size,
-    Filetypes,
-}
-
-const COLUMN_OPTIONS: &[(ReplacedColumn, &str)] = &[
-    (ReplacedColumn::Authors, "Authors"),
-    (ReplacedColumn::Narrators, "Narrators"),
-    (ReplacedColumn::Series, "Series"),
-    (ReplacedColumn::Language, "Language"),
-    (ReplacedColumn::Size, "Size"),
-    (ReplacedColumn::Filetypes, "Filetypes"),
-];
-
 #[component]
 pub fn ReplacedPage() -> Element {
     let _route: crate::app::Route = use_route();
@@ -589,7 +144,6 @@ pub fn ReplacedPage() -> Element {
             filters.read().clone(),
             Some(*from.read()),
             Some(*page_size.read()),
-            *show.read(),
         )
         .await
     })
@@ -618,12 +172,16 @@ pub fn ReplacedPage() -> Element {
             let mut from = from;
             let mut page_size = page_size;
             let mut show = show;
+            let mut selected = selected;
+            let mut last_selected_idx = last_selected_idx;
             sort.set(route_state.sort);
             asc.set(route_state.asc);
             filters_signal.set(route_state.filters);
             from.set(route_state.from);
             page_size.set(route_state.page_size);
             show.set(route_state.show);
+            selected.set(BTreeSet::new());
+            last_selected_idx.set(None);
             last_request_key.set(route_request_key);
             if let Some(resource) = replaced_data.as_mut() {
                 resource.restart();
@@ -631,12 +189,15 @@ pub fn ReplacedPage() -> Element {
         }
     }
 
-    if let Some(value) = &value {
-        let value = value.read();
-        if let Some(Ok(data)) = &*value {
-            cached.set(Some(data.clone()));
+    let cache_value = value;
+    use_effect(move || {
+        if let Some(value) = &cache_value {
+            let value = value.read();
+            if let Some(Ok(data)) = &*value {
+                cached.set(Some(data.clone()));
+            }
         }
-    }
+    });
 
     let data_to_show = {
         if let Some(value) = &value {
