+++ mlm_web_dioxus/src/torrents/mod.rs
@@ -1,0 +1,9 @@
+mod components;
+mod query;
+mod server_fns;
+mod types;
+
+pub use components::TorrentsPage;
+pub use server_fns::{apply_torrents_action, get_torrents_data};
+pub(crate) use types::{TorrentsBulkAction, TorrentsPageColumns, TorrentsRow};
+pub use types::{TorrentsData, TorrentsPageFilter, TorrentsPageSort};
